requirements.txt
openai>=1.0.0
hyperon>=0.2.6
uagents>=0.22.5
uagents-core>=0.3.5
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.10
yfinance>=0.2.32